requirements.txt
asgiref==3.5.2
cssselect==1.1.0
Django==4.0.5
django-environ==0.9.0
django-taggit==3.0.0
django-tinymce==3.4.0
lxml==4.9.0
markdown2==2.4.3
Pillow==9.1.1
pyquery==1.4.3
readtime==2.0.0
sqlparse==0.4.2
tzdata==2022.1
